requirements.txt
glfw==1.7.0
multipledispatch==0.6.0
numpy==1.15.3
Pillow==5.3.0
PyOpenGL==3.1.0
pyrr==0.9.2
PyWavefront==1.0.2
six==1.11.0
